requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.11.1
boto3==1.24.38
botocore==1.27.38
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-appconf==1.0.5
django-imagekit==4.1.0
django-storages==1.12.3
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
lxml==4.9.1
pilkit==2.0
Pillow==9.2.0
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
